rename : Use Layout inherited FlipCardLayout instead of FlipCard
